+++ src/rpi/device.rs
@@ -81,7 +81,7 @@ impl Device {
         }
     }
 
-    pub fn sensor_count(&self) -> usize {
+    pub fn sensor_count(&self) -> u32 {
         match &self.config.model {
             config::Type::BMP085 { .. } => 2,
             config::Type::MCP9808 { .. } => 1,
@@ -89,7 +89,7 @@ impl Device {
         }
     }
 
-    pub fn boolean_count(&self) -> usize {
+    pub fn boolean_count(&self) -> u32 {
         match self.config.model {
             config::Type::BMP085 { .. } => 0,
             config::Type::MCP9808 { .. } => 0,
@@ -97,12 +97,12 @@ impl Device {
         }
     }
 
-    pub fn read_boolean(&self, index: usize) -> Result<bool> {
+    pub fn read_boolean(&self, index: u32) -> Result<bool> {
         match &self.config.model {
-            config::Type::BMP085 { .. } => Err(Error::OutOfBounds(index)),
-            config::Type::MCP9808 { .. } => Err(Error::OutOfBounds(index)),
+            config::Type::BMP085 { .. } => Err(Error::OutOfBounds(index as usize)),
+            config::Type::MCP9808 { .. } => Err(Error::OutOfBounds(index as usize)),
             config::Type::MCP23017 { address, .. } => {
-                let (bank, pin) = mcp23017::index_to_bank_pin(index);
+                let (bank, pin) = mcp23017::index_to_bank_pin(index as usize);
                 let pin = self
                     .mcp23017_state
                     .get_pin(*address, bank, pin, &self.rapi)?;
@@ -111,7 +111,7 @@ impl Device {
         }
     }
 
-    pub fn read_sensor(&self, index: usize) -> Result<(f64, config::Unit)> {
+    pub fn read_sensor(&self, index: u32) -> Result<(f64, config::Unit)> {
         match &self.config.model {
             config::Type::BMP085 { address, mode } => match index {
                 0 => {
@@ -124,35 +124,35 @@ impl Device {
                         .pressure_kpa(*address, *mode, &self.rapi)?;
                     Ok((v as f64, config::Unit::KPa))
                 }
-                _ => Err(Error::OutOfBounds(index)),
+                _ => Err(Error::OutOfBounds(index as usize)),
             },
             config::Type::MCP9808 { address } => match index {
                 0 => {
                     let temp = mcp9808::read_temp(&self.rapi, *address)?;
                     Ok((temp as f64, config::Unit::DegC))
                 }
-                _ => Err(Error::OutOfBounds(index)),
+                _ => Err(Error::OutOfBounds(index as usize)),
             },
-            config::Type::MCP23017 { .. } => Err(Error::OutOfBounds(index)),
+            config::Type::MCP23017 { .. } => Err(Error::OutOfBounds(index as usize)),
         }
     }
 
-    pub fn write_boolean(&mut self, index: usize, value: bool) -> Result<()> {
+    pub fn write_boolean(&mut self, index: u32, value: bool) -> Result<()> {
         match &self.config.model {
-            config::Type::BMP085 { .. } => Err(Error::OutOfBounds(index)),
-            config::Type::MCP9808 { .. } => Err(Error::OutOfBounds(index)),
+            config::Type::BMP085 { .. } => Err(Error::OutOfBounds(index as usize)),
+            config::Type::MCP9808 { .. } => Err(Error::OutOfBounds(index as usize)),
             config::Type::MCP23017 {
                 address,
                 pin_direction,
             } => {
-                let (bank, pin) = mcp23017::index_to_bank_pin(index);
+                let (bank, pin) = mcp23017::index_to_bank_pin(index as usize);
                 let old_dir = self.mcp23017_state.get_pin_direction(bank, pin);
-                if old_dir != pin_direction[index] {
+                if old_dir != pin_direction[index as usize] {
                     self.mcp23017_state.set_pin_direction(
                         *address,
                         bank,
                         pin,
-                        pin_direction[index],
+                        pin_direction[index as usize],
                         &self.rapi,
                     )?;
                 }
